refactor(core): 共通テストヘルパーへの MockLocalizationsDelegate の抽出と docs/testing.md の修正
- 重複していた MockLocalizationsDelegate と MockAppLocalizations を新しく作成した共通ヘルパー widgets_test_helper.dart に抽出
- not_found_screen_test, not_found_screen_golden_test, version_up_dialog_golden_test から重複コードを削除し共通ヘルパーをインポートするようにリファクタリング
- docs/testing.md のゴールデンテスト対象画面の合計数を15箇所に修正し、グラフ画面等の記載漏れを追加
