+++ docs/testing.md
@@ -24,12 +24,12 @@
 
 ### 対象の画面
 
-プロジェクト内のすべての主要な画面や共通ダイアログ（合計11箇所）にゴールデンテストを導入しています。
+プロジェクト内のすべての主要な画面や共通ダイアログ（合計15箇所）にゴールデンテストを導入しています。
 
-- **メイン機能**: ホーム画面、メモ画面、AIチャット画面、設定画面
+- **メイン機能**: ホーム画面、メモ画面、AIチャット画面、グラフ表示画面、グラフ入力画面、設定画面
 - **ユーザー一覧**: ユーザー一覧画面
-- **認証**: ログイン画面、新規登録画面、メール認証画面、パスワード再設定画面
-- **その他**: 404エラー画面、バージョンアップ通知ダイアログ
+- **認証**: Firebaseログイン画面、新規登録画面、メール認証画面、パスワード再設定画面、APIログイン画面
+- **その他**: スプラッシュ画面、404エラー画面、バージョンアップ通知ダイアログ
 
 ---
