+++ test/src/core/widgets/not_found_screen_golden_test.dart
@@ -1,7 +1,6 @@
 import 'package:alchemist/alchemist.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:flutter_sample/l10n/app_localizations.dart';
 import 'package:flutter_sample/src/core/config/app_theme.dart';
 import 'package:flutter_sample/src/core/widgets/not_found_screen.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -10,19 +9,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:mocktail/mocktail.dart';
 
 import 'not_found_screen_test.dart';
-
-// モック化された多言語デリゲートクラスを定義します
-class MockLocalizationsDelegate
-    extends LocalizationsDelegate<AppLocalizations> {
-  const MockLocalizationsDelegate(this.mock);
-  final MockAppLocalizations mock;
-  @override
-  bool isSupported(Locale locale) => true;
-  @override
-  Future<AppLocalizations> load(Locale locale) async => mock;
-  @override
-  bool shouldReload(covariant _) => false;
-}
+import 'widgets_test_helper.dart';
 
 void main() {
   group('NotFoundScreen Golden Tests', () {
