+++ test/src/core/widgets/version_up_dialog_golden_test.dart
@@ -1,25 +1,12 @@
 import 'package:alchemist/alchemist.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:flutter_sample/l10n/app_localizations.dart';
 import 'package:flutter_sample/src/core/config/app_theme.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:mocktail/mocktail.dart';
 
-class MockAppLocalizations extends Mock implements AppLocalizations {}
-
-class MockLocalizationsDelegate
-    extends LocalizationsDelegate<AppLocalizations> {
-  const MockLocalizationsDelegate(this.mock);
-  final MockAppLocalizations mock;
-  @override
-  bool isSupported(Locale locale) => true;
-  @override
-  Future<AppLocalizations> load(Locale locale) async => mock;
-  @override
-  bool shouldReload(covariant _) => false;
-}
+import 'widgets_test_helper.dart';
 
 void main() {
   group('VersionUpDialog Golden Tests', () {
