+++ test/src/core/widgets/widgets_test_helper.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_sample/l10n/app_localizations.dart';
+import 'package:mocktail/mocktail.dart';
+
+// テスト用の多言語翻訳のモックです
+class MockAppLocalizations extends Mock implements AppLocalizations {}
+
+// 翻訳データを読み込むためのモックデリゲートです
+class MockLocalizationsDelegate
+    extends LocalizationsDelegate<AppLocalizations> {
+  const MockLocalizationsDelegate(this.mock);
+  final MockAppLocalizations mock;
+  @override
+  bool isSupported(Locale locale) => true;
+  @override
+  Future<AppLocalizations> load(Locale locale) async => mock;
+  @override
+  bool shouldReload(covariant _) => false;
+}
